Put validations out of the incident controllers file
