+++ app/models/ti.py
@@ -7,7 +7,9 @@ users = []
 
 class Ti:
     def __init__(self, *args):
-
+        """This method acts as a constructor
+            for the users class, it initialises all class attributes
+        """
         self.firstname = args[0]
         self.lastname = args[1]
         self.othernames = args[2]
@@ -18,6 +20,8 @@ class Ti:
         self.password = args[7]
 
     def get_dictionary(self):
+        """this method converts the class attributes into json objects
+        """
         return{
             "user_id": len(users) + 1,
             "firstname": self.firstname,
