+++ app/validators.py
@@ -1,36 +1,71 @@
-# # def validate_input(createdby, location):
+import re
 
-# # zif not createdby or createdby.isspace():
-# #             return 'please enter the id of the creator of this redflag'
-# #         elif not location or location.isspace():
-# #             return 'Enter location.'
 
-# from flask import Flask, jsonify, make_response, request
-# def validate_data(value):
-#     if not value or value.isspace() or len(value)<2 or len(value)>24 or type(value) == int:
-#         return jsonify({"status":404, "message":"please enter all items in the right formart"}), 404
-# # def validate_data_type(username):
-# #     if not type(username) == str:
-# #         return 'status must be an String!!'
-# def validate_datas(value):
-#     if not value or value.isspace() or len(value)<2 or len(value)>24 or type(value) == int:
-#         return jsonify({"status":404, "message":"please enter all items in the right formart"}), 404
+class Validators:
 
-# def validate_intdata_type(numbers):
-#     if not type(numbers) == int:
-#         return 'status must be a number!!'
+    def validate_location(self, location):
+        """Method where all validations are done"""
+        geo_location = re.compile(
+            "^[0-9]{2}(.)[0-9]{2}( )[0-9]{2}(.)[0-9]{2}$")
 
-# def validate_keys(value, lst):
-#     if value not in lst:
-#         return jsonify({"status":404, "message":"All fields must be present"}), 404
+        if not location or location.isspace():
+            return 'please enter the location of this redflag'
 
-# def input_jsonformat(data):
-#     if not type(data) == dict:
-#         return jsonify({
-#             "message":'Data must be in dictionary format',
-#             "required format":{"createdby":2,"incidenttype":"jgjjhghhgjh","location":"hjgjhg", "status":"jhgkjh", "image":"hjgjhg", "video":"vhgfghf", "comment":"hjgjhkj"}
-#             }), 400
+        elif not geo_location.match(location):
+            string1 = 'invalid location, please enter the lat, long cordinates'
+            string2 = 'in this formant, 25.22 56.22'
+            return string1 + string2
 
-# # def validate_status(status):
-# #     if not status=="draft" or status=="resolved":
-# #         return "invalid input"
+    def validate_input(self, *args):
+        """Method where all validations are done"""
+        createdby = args[0]
+        incident_type = args[1]
+        status = args[2]
+
+        if not createdby:
+            return 'please enter the id of the creator of this redflag'
+        elif not isinstance(createdby, int):
+            return 'createdby should be an id of the creator of the redflag'
+        elif not incident_type or incident_type.isspace():
+            return 'Enter a redflag.'
+        elif incident_type != "redflag" and incident_type != "intervention":
+            return 'Incident type should either be a redflag or intervention.'
+        elif status != "draft":
+            return 'status should either be draft,\
+                underinvestigation, resolved or rejected'
+
+    def validate_coment(self, comment):
+        if not comment or comment.isspace():
+            return 'Enter the comment'
+        if isinstance(comment, int):
+            return 'Comment should be a string'
+
+    def validate_user_details(self, *args):
+        firstname = args[0]
+        lastname = args[1]
+        email = args[2]
+        username = args[3]
+        password = args[4]
+        phoneNumber = args[5]
+        isAdmin = args[6]
+        othernames = args[7]
+
+        email_validation = re.compile(
+            "(^[a-zA-z0-9_.]+@[a-zA-z0-9-]+\.[a-z]+$)")
+
+        if not isinstance(firstname, str) or not isinstance(lastname, str)\
+                or not isinstance(username, str) or \
+                not isinstance(othernames, str):
+            return "Invalid input, Enter a string"
+        elif not firstname or firstname.isspace() or not lastname or\
+                lastname.isspace() or not email or email.isspace() or \
+                not username or username.isspace() or not password or \
+                password.isspace()or not isAdmin or isAdmin.isspace():
+            return "Enter all items"
+        elif type(phoneNumber) == str:
+            return "The phone number should be an integer"
+
+        elif isAdmin != 'false' and isAdmin != 'true':
+            return "IsAdmin should either be true or false"
+        elif not email_validation.match(email):
+            return 'Invalid email, it should be in this format; kals@gma.com'
